+++ lab2/DebugPractice/DebugExercise3.java
@@ -1,16 +1,17 @@
 /**
  * Created by jug on 1/22/18.
+ * Debugged by Xiaoyu on 5/23/18
  */
 public class DebugExercise3 {
-    public static int countTurnips(In in) {
-        int totalTurnips = 0;
+    public static long countTurnips(In in) {
+        long totalTurnips = 0;
         while (!in.isEmpty()) {
             String vendor = in.readString();
             String foodType = in.readString();
             double cost = in.readDouble();
-            int numAvailable = in.readInt();
+            long numAvailable = in.readInt();
             if (foodType.equals("turnip")) {
-                int newTotal = totalTurnips + numAvailable;
+                long newTotal = totalTurnips + numAvailable;
                 totalTurnips = newTotal;
             }
             in.readLine();
